Refactor preprocessing processors and logging
