+++ src/sp500_analysis/application/preprocessing/processors/eoe.py
@@ -1,0 +1,14 @@
+from pathlib import Path
+from sp500_analysis.shared.logging.logger import configurar_logging
+
+
+class EOEProcessor:
+    def __init__(self, data_root='data/0_raw', log_file='eoe_universal_processor.log'):
+        self.data_root = data_root
+        self.logger = configurar_logging(log_file, name='EOEProcessor')
+
+    def run(self, output_file):
+        """Placeholder run implementation for EOE processing."""
+        self.logger.info('Running EOE processor')
+        Path(output_file).touch()
+        return True
