+++ src/sp500_analysis/application/preprocessing/processors/fred.py
@@ -1,0 +1,15 @@
+from pathlib import Path
+from sp500_analysis.shared.logging.logger import configurar_logging
+
+
+class FredProcessor:
+    def __init__(self, config_file, data_root='data/0_raw', log_file='freddataprocessor.log'):
+        self.config_file = config_file
+        self.data_root = data_root
+        self.logger = configurar_logging(log_file, name='FredProcessor')
+
+    def run(self, output_file):
+        """Placeholder run implementation for FRED data processing."""
+        self.logger.info('Running FRED processor')
+        Path(output_file).touch()
+        return True
